% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{One Sample Bootstrap}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{sample}

\item{conf.level}{1-alpha}

\item{iter}{number of iterations}

\item{mu0}{Null Hypothesis value}

\item{test}{}
}
\value{

}
\description{
One Sample Bootstrap
}
