% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire distance and damage.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{distance}{distance a fire is from a fire station, in miles}
  \item{damage}{x 1000 dollars in damage}
  ...
}
}
\source{
\url{http://www.crcpress.com/Statistics-for-Engineering-and-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A dataset containing the distance a fire is from a fire station and how much damge it caused.
}
\keyword{datasets}
