% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{A Binomial Sampling Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Length}

\item{p}{Probability}
}
\value{
A plot
}
\description{
A Binomial Sampling Function
}
\examples{
\dontrun{mybin(iter=100,n=10, p=0.5)}
}
