% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My Central Limit Theorem Function}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{distribution size}

\item{iter}{number of trials}

\item{lambda}{expected value}

\item{...}{}
}
\value{

}
\description{
My Central Limit Theorem Function
}
\examples{
\dontrun{mycltp(n=2,iter=10000, lambda = 4)}
}
