% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{data}

\item{mu}{mean}

\item{sig}{standard deviation}

\item{...}{}
}
\value{

}
\description{
mymlnorm
}
